pyproject.toml
[project]
name = "agentic-with-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.11.4",
    "duckduckgo-search>=8.1.1",
    "google-genai==1.66.0",
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.1.2",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
